feat: 피드 v1 페이지 제거, AI 채팅 페이지 추가
